Navmesh setup and ready for isometric cam
So I've set up the navmesh on the player and the scripts to point and click from the main camera on the environment. However, I'm not sure how to make the isometric view and I had to remove all the camera scripts from the other scene. Let me know if you can get the isometric working for this guy.